Refactor the chat logic
